requirements.txt
requests>=2.10.0
pyyaml>=3.12
six>=1.10
